# [CUDA] Enable CUDA Graph on CUDA Toolkit < 12.x