@gaugarg-nv 2025-03-17
> Seems to cause error in the MUSA build:
> 
> https://github.com/ggml-org/llama.cpp/actions/runs/13860248342/job/38872230770?pr=12394#step:6:6724

Added a change that should fix issues with MUSA build.